+++ t-tabular/src/main/java/eu/unifiedviews/plugins/transformer/tabular/TabularVaadinDialog.java
@@ -17,13 +17,12 @@ import com.vaadin.ui.TabSheet.Tab;
 import cz.cuni.mff.xrg.uv.transformer.tabular.Tabular;
 import cz.cuni.mff.xrg.uv.transformer.tabular.TabularConfig_V2;
 import cz.cuni.mff.xrg.uv.transformer.tabular.TabularOntology;
-import cz.cuni.mff.xrg.uv.transformer.tabular.TabularConfig_V2.AdvanceMapping;
 import cz.cuni.mff.xrg.uv.transformer.tabular.column.ColumnInfo_V1;
+import cz.cuni.mff.xrg.uv.transformer.tabular.column.NamedCell_V1;
 import cz.cuni.mff.xrg.uv.transformer.tabular.parser.ParserType;
 import cz.cuni.mff.xrg.uv.transformer.tabular.parser.ParserXls;
 import eu.unifiedviews.dpu.config.DPUConfigException;
 import eu.unifiedviews.helpers.dpu.vaadin.dialog.AbstractDialog;
-import eu.unifiedviews.plugins.transformer.tabular.column.NamedCell_V1;
 import eu.unifiedviews.plugins.transformer.tabular.gui.PropertyGroup;
 import eu.unifiedviews.plugins.transformer.tabular.gui.PropertyGroupAdv;
 import eu.unifiedviews.plugins.transformer.tabular.gui.PropertyNamedCell;
@@ -62,7 +61,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
     private CheckBox checkGenerateTableClass;
 
-    private CheckBox checkGenerateLabels ;
+    private CheckBox checkGenerateLabels;
 
     private TextField txtCsvQuoteChar;
 
@@ -79,6 +78,8 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
     private CheckBox checkXlsHasHeader;
 
     private CheckBox checkXlsStripHeader;
+
+    private CheckBox checkDbfTrimString;
 
     /**
      * Layout for basic column mapping.
@@ -143,7 +144,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         this.txtBaseUri.setRequiredError("Resource URI base must be supplied.");
         this.txtBaseUri.setDescription(
                 "This value is used as base URI for automatic column property generation "
-                + "and also to create absolute URI if relative uri is provided in 'Property URI' column.");
+                        + "and also to create absolute URI if relative uri is provided in 'Property URI' column.");
         generalLayout.addComponent(this.txtBaseUri);
 
         this.txtKeyColumnName = new TextField("Key column");
@@ -198,7 +199,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         this.checkStaticRowCounter = new CheckBox("Use static row counter");
         this.checkStaticRowCounter.setDescription(
                 "If checked and multiple files are precessed, then those files share the same row counter."
-                + "The process can be viewsed as if files are appended before parsing.");
+                        + "The process can be viewsed as if files are appended before parsing.");
         checkLayout.addComponent(checkStaticRowCounter);
 
         this.checkAdvancedKeyColumn = new CheckBox("Advanced key column");
@@ -214,8 +215,8 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         this.checkTableSubject = new CheckBox("Generate subject for table");
         this.checkTableSubject.setDescription(
                 "If checked then a subject for each table that point to all rows in given table is created. "
-                + "Used predicate is '" + TabularOntology.TABLE_HAS_ROW + "'. By predicate '" + TabularOntology.TABLE_SYMBOLIC_NAME + "'."
-                + "Symbolic name of source file is also attached.");
+                        + "Used predicate is '" + TabularOntology.TABLE_HAS_ROW + "'. By predicate '" + TabularOntology.TABLE_SYMBOLIC_NAME + "'."
+                        + "Symbolic name of source file is also attached.");
         checkLayout.addComponent(this.checkTableSubject);
 
         this.checkAutoAsString = new CheckBox("Auto type as string");
@@ -302,6 +303,19 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
             }
         });
 
+        // DBF
+        final FormLayout dbfLayout = new FormLayout();
+        dbfLayout.setImmediate(true);
+        dbfLayout.setSpacing(true);
+        dbfLayout.setWidth("100%");
+        dbfLayout.setHeight("-1px");
+        dbfLayout.addComponent(new Label("DBF specific settings"));
+
+        this.checkDbfTrimString = new CheckBox("Remove trailing spaces");
+        this.checkDbfTrimString.setDescription("If checked then for every loaded string the leading and "
+                + "trailing spaces are removed before the value is futher processed.");
+        dbfLayout.addComponent(this.checkDbfTrimString);
+
         // --------------------- Mapping - simple ---------------------
         this.basicLayout = new GridLayout(5, 1);
         this.basicLayout.setWidth("100%");
@@ -369,8 +383,8 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         Tab tabAdv = propertiesTab.addTab(this.advancedLayout, "Advanced - experimental functionality!");
         tabAdv.setDescription(
                 "Templates based on http://w3c.github.io/csvw/csv2rdf/#. If { or } is part of column name"
-                + "then before use they must be escaped ie. \\{ or \\} should be used."
-                + "Use \"...\" to denote literal and <...>  to denote uri. '...' then represent the content of literal/uri.");
+                        + "then before use they must be escaped ie. \\{ or \\} should be used."
+                        + "Use \"...\" to denote literal and <...>  to denote uri. '...' then represent the content of literal/uri.");
         Tab tabXls = propertiesTab.addTab(this.xlsStaticLayout, "Xls mapping");
         tabXls.setDescription(
                 "Can be used for static mapping of cells to named cells. Named cells are accesible as extension in every row.");
@@ -378,12 +392,13 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         // -------------------------------------------------------------
         // top layout with configuration
         final HorizontalLayout configLayout = new HorizontalLayout();
-        configLayout.setWidth("100%");
+        configLayout.setWidth("-1px");
         configLayout.setHeight("-1px");
         configLayout.setSpacing(true);
 
         configLayout.addComponent(generalLayout);
         configLayout.addComponent(csvLayout);
+        configLayout.addComponent(dbfLayout);
         configLayout.addComponent(xlsLayout);
 
         // main layout for whole dialog
@@ -484,12 +499,13 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
     /**
      * Based on given type update properties.
-     *
+     * 
      * @param value
      */
     private void setControllStates(ParserType value) {
         boolean csvEnabled = value == ParserType.CSV;
         boolean xlsEnabled = value == ParserType.XLS;
+        boolean dbfEnabled = value == ParserType.DBF;
 
         txtCsvQuoteChar.setEnabled(csvEnabled);
         txtCsvDelimeterChar.setEnabled(csvEnabled);
@@ -504,11 +520,13 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         for (PropertyNamedCell namedCell : xlsNamedCells) {
             namedCell.setEnabled(xlsEnabled);
         }
+
+        checkDbfTrimString.setEnabled(dbfEnabled);
     }
 
     /**
      * Add new line (component) into tab "Simple" mapping.
-     *
+     * 
      * @param name
      * @param setting
      */
@@ -523,7 +541,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
     /**
      * Add new line (component) into tab "Advanced" mapping.
-     *
+     * 
      * @param uri
      * @param template
      */
@@ -538,7 +556,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
     /**
      * Add new line into "xls" mapping.
-     *
+     * 
      * @param name
      * @param column
      * @param row
@@ -554,7 +572,7 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
     /**
      * Parse given string and use it to set column (properties) names. Original data are lost.
-     *
+     * 
      * @param str
      */
     private void importColumnNames(String str, String separator) {
@@ -599,9 +617,12 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
         // save uri
         //
         String uriStr = c.getBaseURI();
-        if (!uriStr.endsWith("/")) {
+
+        if (!uriStr.endsWith("/") && !uriStr.endsWith("#")) {
+            //uri does not end with "/" or "#", add "/" automatically
             uriStr = uriStr + "/";
         }
+
         try {
             new java.net.URI(uriStr);
         } catch (URISyntaxException ex) {
@@ -639,6 +660,11 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
             loadCellMapping(Collections.EMPTY_LIST);
             checkXlsHasHeader.setValue(true);
             checkXlsStripHeader.setValue(false);
+        }
+        if (c.getTableType() == ParserType.DBF) {
+            checkDbfTrimString.setValue(c.isDbfTrimString());
+        } else {
+            checkDbfTrimString.setValue(false);
         }
         //
         // other data
@@ -734,6 +760,8 @@ public class TabularVaadinDialog extends AbstractDialog<TabularConfig_V2> {
 
             cnf.setHasHeader(checkXlsHasHeader.getValue());
             cnf.setStripHeader(checkXlsStripHeader.getValue());
+        } else if (value == ParserType.DBF) {
+            cnf.setDbfTrimString(checkDbfTrimString.getValue());
         }
         //
         // other data