+++ t-gzipper/CHANGELOG.md
@@ -1,10 +1,19 @@
 T-Gzipper
 ----------
 
-v1.0.1-SNAPSHOT
+v1.0.3
+---
+* Update to helpers 2.1.3
+
+v1.0.2
+---
+* Fixed documentation (About)
+* Update to API 2.1.2 (properly displayed about tab in UTF-8 encoding)
+
+v1.0.1
 ---
 * Improved description (About)
 
-v1.0.0-SNAPSHOT
+v1.0.0
 ---
 * initial release